Fix NPE

+++ plugins/org.eclipse.ease.lang.python.py4j/src/org/eclipse/ease/lang/python/py4j/internal/Py4jScriptEngine.java
@@ -218,7 +218,9 @@ public class Py4jScriptEngine extends AbstractScriptEngine {
 		// to cleanup.
 		// XXX: This is an issue solved by PyDev, resolving it here fully is not
 		// the logical course of action.
-		fPythonProcess.destroyForcibly();
+		if (fPythonProcess != null) {
+			fPythonProcess.destroyForcibly();
+		}
 	}
 
 	@Override
